Created a wrapper for RequestReportRequest to return to clients, that...
also contains additional information about the request object, like the Frequency with which amazon updates information for that report. This can have a direct impact on reducing the number of requests clients make to amazon the reporting endpoints (if a report is updated daily, there's no need for it to be fetched 3 times a day). Also added ContentUpdateFrequency enum with some generic values