Added extra client configuration options for the retry of the amazon
SubmitFeed operation. This is because the reason for which SubmitFeed operation might fail could be very different from the reason for which the feed processing step might be failing. (for example network issues of a feed in invalid xml format). Refactored the logic for calculating if the retry period has elapsed into a helper class.