Executing SaveChanges after every amazon request
Even though this might seem counter-intuitive, calling save changes after every amazon call will vastly improve the efficiency of the Poll methods, because 1. the calls to amazon are syncronous. and 2. some calls to amazon are resolved immediately - no retry has to be performed. This way, it is possible to perform multiple operations from the pipeline of the same report/feed within the same Poll() call. As a result the Poll method will be called a smaller amount of times in order to resolve a given number of requests.