Replaced static property EasyMwsOptions.Defaults with static method..
that always returns a new copy of the default options.
If an easyMwsClient user would have initialised his easyMwsOptions with the defaults, when he would change any value, he would actually modify the a value on the static Defaults property ref. Given that another client with different credentials/settings would be initialised on the same machine, an attempt to get the default options, would actually return the set of defaults modified by the first user. xD (also fixed some unit tests vodoo that happened because of the described scenario)