Added FeedSubmissionCallback extension GetRegionAndTypeString
that can be used to obtain a friendly string uniquely identifying an entity (feed region and feed type). Also created RegionAndTypeComputed property that will try to deserialize FeedSubmissionData in order to obtain ReportType - the first time the getter is called, and caching the value for later use. Given thet FeedSubmissionData should only be set once on an entity (no use for it ever getting modified) - this seems an acceptable tradeoff. also made small adjustments to the same functionality on ReportRequestCallback to stay consistent