Added connection string option in EasyMwsOptions
By default this option is null, and the connection string will be obtained from a config file depending on the platform calling the EasyMws code.
If this option is set to a non-nullOrEmpty value, then this provided connection string will bypass the logic trying to read the connection string from a configuration file.