Remove event invoking client implementation, as it is not viable
in the same solution - this is due to the fact that Feeds/Reports that would be queued one client type (Callback/Event based) have to be invoked back using the same expected invocation strategy - and this would have to be handled by whoever is using the client, adding an extra layer of complexity. - therefore the implementation moves away from the initial purpose of the package (Easy mws).