Add IsLocked column on entry entities
The column marks whether the table row is currently being handled by an EasyMws process. The value is 'true' if a process grabbed the row and is currently working/waiting to perform logic on it. The value is marked as 'false' once a process has finished doing changes to that row.