Fix memory leaks
These were caused for some reason because of using Expression<Func<object, bool>> to pass filters to custom where / any or similar internal service methods.
The reason seems to be a pile up of evaluation expressions in memory thanks to Expression<Func<>> evaluating to IQueryable. After Expression<Func<>> was replaced with Func<> it evaluates to IEnumerable, so evaluation expressions don't keep piling up in memory. The downside is queries now get all db entries in memory, but this should not be a problem because the cleanup mechanism prevents entry tables from getting too big (entries older than 1/2 days are deleted). To put things in context, the memory leaks in question could lead to OutOfMemory exceptions popping up on a machine.